requirements.txt
Flask==2.3.2
g4f==0.1.9.4
openai==0.28.0
python_bcrypt==0.3.2
Requests==2.31.0
supabase==2.3.0
gunicorn